requirements.txt
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==37.0.2
distlib==0.3.4
filelock==3.6.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
jeepney==0.8.0
keyring==23.5.0
lockfile==0.12.2
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.2
poetry-core==1.0.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
requests==2.27.1
requests-toolbelt==0.9.1
SecretStorage==3.3.2
shellingham==1.4.0
six==1.16.0
tomlkit==0.10.2
urllib3==1.26.9
utils==1.0.1
virtualenv==20.14.1
webencodings==0.5.1
zipp==3.8.0
jinja2==3.1.2